Support for using multiple cores and external files
Added the `--cores` option, which in local mode, the number of CPU cores available to run rules and in Slurm mode, the number of sbatch jobs submitted in parallel. 
Added an `ext` directory with a subfolder for scripts. Instructions on how to use it added to 'Extending the Module' (module README) and 'Writing Pipeline Steps (Rules)' (template README). Explicitly for external resources that cannot easily be integrated into either `utils.py` or `parameters.yaml`, and are not as large as databases that would justify an externally stored download.
Added an example of `workflow_mode` based on whether Nanopore vs. Illumina and bin refinement vs. binning algorithms alone are expected so that users can see multiple layers of output toggling, and put explanation to both the template and module READMEs.